requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
PyPDF2==3.0.1
httpx==0.27.0
python-multipart==0.0.9
gunicorn==22.0.0
